agent-infra-docker/pyproject.toml
[project]
name = "agent-infra-docker"
version = "1.0.0"
requires-python = ">=3.11,<4"
readme = "README.md"
license-files = ["LICENSE"]
authors = [{ name = "Agno", email = "hello@agno.com" }]

dependencies = [
  "agno==2.0.4",
  "agno-infra==1.0.2",
  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "certifi==2025.7.14",
  "charset-normalizer==3.4.3",
  "click==8.2.1",
  "ddgs==9.5.5",
  "distro==1.9.0",
  "dnspython==2.7.0",
  "docstring-parser==0.17.0",
  "email-validator==2.2.0",
  "fastapi-cli==0.0.8",
  "fastapi-cloud-cli==0.1.5",
  "fastapi[standard]==0.116.1",
  "gitdb==4.0.12",
  "gitpython==3.1.45",
  "google-genai>=1.38.0",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httptools==0.6.4",
  "httpx==0.28.1",
  "idna==3.10",
  "jinja2==3.1.6",
  "jiter==0.10.0",
  "lxml==6.0.0",
  "markdown-it-py==3.0.0",
  "markupsafe==3.0.2",
  "mdurl==0.1.2",
  "numpy==2.3.2",
  "openai==1.97.1",
  "packaging==25.0",
  "pgvector==0.4.1",
  "primp==0.15.0",
  "psycopg-binary==3.2.9",
  "psycopg[binary]==3.2.9",
  "pydantic==2.11.7",
  "pydantic-core==2.33.2",
  "pydantic-settings==2.10.1",
  "pygments==2.19.2",
  "python-dotenv==1.1.1",
  "python-multipart==0.0.20",
  "pyyaml==6.0.2",
  "requests==2.32.5",
  "rich==14.1.0",
  "rich-toolkit==0.14.9",
  "rignore==0.6.4",
  "sentry-sdk==2.33.2",
  "shellingham==1.5.4",
  "smmap==5.0.2",
  "sniffio==1.3.1",
  "sqlalchemy==2.0.41",
  "starlette==0.47.2",
  "tomli==2.2.1",
  "tqdm==4.67.1",
  "typer==0.16.0",
  "typing-extensions==4.14.1",
  "typing-inspection==0.4.1",
  "urllib3==2.5.0",
  "uvicorn==0.35.0",
  "uvloop==0.21.0",
  "watchfiles==1.1.0",
  "websockets==15.0.1",
  "pymupdf>=1.23.0",
  "pandas>=2.2.0",
  "openpyxl>=3.1.0",
  "pytesseract>=0.3.10",
  "Pillow>=10.0.0",
  "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = ["mypy", "ruff", "pytest", "pytest-asyncio"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

[tool.ruff]
line-length = 120
exclude = [".venv*"]
[tool.ruff.lint.per-file-ignores]
# Ignore `F401` (import violations) in all `__init__.py` files
"__init__.py" = ["F401", "F403"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = [".venv*"]

[[tool.mypy.overrides]]
module = ["pgvector.*", "setuptools.*", "nest_asyncio.*", "agno.*", "requests.*"]
ignore_missing_imports = true

[tool.uv.pip]
no-annotate = true
